Add the TagSelection component to the editable OrgCard, too
Since the new tags are associated with email addresses, they should be
available in the edit view of the OrgCard, too. It's not quite enough to
add it to the editable branch of the HTML in ContactEmail, we also have
to pass the annotationHints into the ContactEmail in the edtiable branch
of the HTML in the OrgCard component.

The styling of the TagSelection component in the editable OrgCard
doesn't fit yet, but it's functional and updates automatically when the
email address is modified.