Rework HTML and styling of the TagSelection component
To make the TagSelection component fit better into the two different
places where it's used, particularly in term of styling, we move some
part of the HTML out of TagSelection into the places in ContactEmail
where TagSelection is used so that it's easier to use different styles
in the two places.

This moves the label out of the TagSelection into ContactEmail, which is
a bit unfortunate, but under the circumstances seems like a good
compromise, because it makes it easy to explicitly apply a style to the
label.