contact: Proof of concept for better search result summary
Instead of showing the summary of the search results in the boxes with
the search input fields, the summary is now shown in a new box above the
search results. The summary also doesn't change anymore when the user
types in the search input fields. In addition to the result counts
('Found X auto-imported ...') it also shows which search criterion was
used.