Experiment to evaluate JSON Schema for validation
OrgCards use a JSON Schema definition for some parts of the organisation
data to validate user inputs. Invalid inputs are highlighted directly in
the form. Using a JSON schema makes it possible to use the same schema
in the client and the server, avoiding code duplication.

This is only a proof of concept that demonstrates the principle. There
are quite a few things missing. For instance, the schema does not cover
all of the organisation data yet and when validation failes, the user
can still commit the changes.