Fix instantiation of national certs
The template has to include the country_code attribute in order for the
JSON schema validation to work correctly (without it, in newly created
national certs the country would not be considered invalid because it's
empty).

In addtion the address attribute and the method name, newFqdn, seem to
stem from incomplete adaptation of copy&pasted code, so both of these
are fixed.